pyproject.toml
[tool.poetry]
name = "mto"
version = "0.1.0"
description = ""
authors = ["Divider <it-cio@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pandas = "^1.4.4"
python-docx = "^0.8.11"
XlsxWriter = "^3.0.3"
python-dotenv = "^0.21.0"
aioschedule = "^0.5.2"
aspose-words = "^22.9.0"
SQLAlchemy = "^1.4.41"
PyMySQL = "^1.0.2"
openpyxl = "^3.0.10"
tqdm = "^4.64.1"
panoramisk = "^1.4"
aiohttp = "^3.8.3"
bs4 = "^0.0.1"
uvicorn = "^0.18.3"
fastapi = "^0.85.0"
PyJWT = "^2.5.0"
passlib = "^1.7.4"
fastapi-crudrouter = "^0.8.5"
cryptography = "^38.0.1"
alembic = "^1.8.1"
bcrypt = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
